+++ packages/rs-dpp/src/identity/mod.rs
@@ -2,6 +2,7 @@ pub use credits_converter::*;
 pub use credits_converter::*;
 pub use get_biggest_possible_identity::*;
 pub use identity::*;
+pub use identity_stub::*;
 pub use identity_facade::*;
 pub use identity_public_key::*;
 
@@ -18,3 +19,4 @@ mod credits_converter;
 pub mod errors;
 pub mod factory;
 pub mod serialize;
+mod identity_stub;
