+++ packages/rs-drive/src/drive/identity/fetch/mod.rs
@@ -29,34 +29,12 @@ use grovedb::{PathQuery, SizedQuery, TransactionArg};
 use integer_encoding::VarInt;
 use std::collections::BTreeMap;
 
-impl Drive {
-    /// Fetches the Identity's balance as an identity stub from the backing store
-    /// Passing apply as false get the estimated cost instead
-    pub fn fetch_identity_stub_with_balance(
-        &self,
-        identity_id: [u8; 32],
-        apply: bool,
-        transaction: TransactionArg,
-    ) -> Result<Option<Identity>, Error> {
-        let mut drive_operations: Vec<DriveOperation> = vec![];
-        Ok(self
-            .fetch_identity_balance_operations(
-                identity_id,
-                apply,
-                transaction,
-                &mut drive_operations,
-            )?
-            .map(|balance| Identity {
-                protocol_version: PROTOCOL_VERSION,
-                id: Identifier::new(identity_id),
-                public_keys: Default::default(),
-                balance,
-                revision: 0,
-                asset_lock_proof: None,
-                metadata: None,
-            }))
-    }
+mod balance;
+mod fetch_by_public_key_hashes;
+mod full_identity;
+mod revision;
 
+impl Drive {
     /// The query for the identity revision
     pub fn identity_revision_query(identity_id: [u8; 32]) -> PathQuery {
         let identity_path = identity_path_vec(identity_id.as_slice());
@@ -70,264 +48,6 @@ impl Drive {
                 offset: None,
             },
         }
-    }
-
-    /// Fetches the Identity's revision from the backing store
-    /// Passing apply as false get the estimated cost instead
-    pub fn fetch_identity_revision(
-        &self,
-        identity_id: [u8; 32],
-        apply: bool,
-        transaction: TransactionArg,
-    ) -> Result<Option<u64>, Error> {
-        let mut drive_operations: Vec<DriveOperation> = vec![];
-        self.fetch_identity_revision_operations(
-            identity_id,
-            apply,
-            transaction,
-            &mut drive_operations,
-        )
-    }
-
-    /// Fetches the Identity's revision from the backing store
-    /// Passing apply as false get the estimated cost instead
-    pub fn fetch_identity_revision_with_fees(
-        &self,
-        identity_id: [u8; 32],
-        block_info: &BlockInfo,
-        apply: bool,
-        transaction: TransactionArg,
-    ) -> Result<(Option<u64>, FeeResult), Error> {
-        let mut drive_operations: Vec<DriveOperation> = vec![];
-        let value = self.fetch_identity_revision_operations(
-            identity_id,
-            apply,
-            transaction,
-            &mut drive_operations,
-        )?;
-        let fees = calculate_fee(None, Some(drive_operations), &block_info.epoch)?;
-        Ok((value, fees))
-    }
-
-    /// Creates the operations to get Identity's revision from the backing store
-    /// This gets operations based on apply flag (stateful vs stateless)
-    pub(crate) fn fetch_identity_revision_operations(
-        &self,
-        identity_id: [u8; 32],
-        apply: bool,
-        transaction: TransactionArg,
-        drive_operations: &mut Vec<DriveOperation>,
-    ) -> Result<Option<u64>, Error> {
-        let direct_query_type = if apply {
-            DirectQueryType::StatefulDirectQuery
-        } else {
-            DirectQueryType::StatelessDirectQuery {
-                in_tree_using_sums: false,
-                query_target: QueryTargetValue(1),
-            }
-        };
-        let identity_path = identity_path(identity_id.as_slice());
-        match self.grove_get_raw(
-            identity_path,
-            &[IdentityTreeRevision as u8],
-            direct_query_type,
-            transaction,
-            drive_operations,
-        ) {
-            Ok(Some(Item(encoded_revision, _))) => {
-                let revision = u64::from_be_bytes(encoded_revision.try_into().map_err(|_| {
-                    Error::Drive(DriveError::CorruptedElementType(
-                        "identity revision was not 8 bytes as expected",
-                    ))
-                })?);
-
-                Ok(Some(revision))
-            }
-
-            Ok(None) => Ok(None),
-
-            Ok(Some(..)) => Err(Error::Drive(DriveError::CorruptedElementType(
-                "identity revision was present but was not identified as an item",
-            ))),
-
-            Err(e) => Err(e),
-        }
-    }
-
-    /// Given an identity, fetches the identity with its flags from storage.
-    pub fn fetch_identity_balance_with_keys(
-        &self,
-        identity_key_request: IdentityKeysRequest,
-        transaction: TransactionArg,
-    ) -> Result<Option<Identity>, Error> {
-        // let's start by getting the balance
-        let id = Identifier::new(identity_key_request.identity_id);
-        let balance =
-            self.fetch_identity_balance(identity_key_request.identity_id, true, transaction)?;
-        if balance.is_none() {
-            return Ok(None);
-        }
-        let balance = balance.unwrap();
-
-        let public_keys = self.fetch_identity_keys::<KeyIDIdentityPublicKeyPairBTreeMap>(
-            identity_key_request,
-            transaction,
-        )?;
-        Ok(Some(Identity {
-            protocol_version: PROTOCOL_VERSION,
-            id,
-            public_keys,
-            balance,
-            revision: u64::MAX,
-            asset_lock_proof: None,
-            metadata: None,
-        }))
-    }
-
-    /// Fetches an identity with all its information and
-    /// the cost it took from storage.
-    pub fn fetch_full_identity_with_costs(
-        &self,
-        identity_id: [u8; 32],
-        epoch: &Epoch,
-        transaction: TransactionArg,
-    ) -> Result<(Option<Identity>, FeeResult), Error> {
-        let mut drive_operations: Vec<DriveOperation> = vec![];
-        let maybe_identity =
-            self.fetch_full_identity_operations(identity_id, transaction, &mut drive_operations)?;
-        let fee = calculate_fee(None, Some(drive_operations), epoch)?;
-        Ok((maybe_identity, fee))
-    }
-
-    /// The query getting all keys and balance and revision
-    pub fn full_identity_query(identity_id: [u8; 32]) -> Result<PathQuery, Error> {
-        let balance_query = Self::identity_balance_query(identity_id);
-        let revision_query = Self::identity_revision_query(identity_id);
-        let key_request = IdentityKeysRequest::new_all_keys_query(identity_id);
-        let all_keys_query = key_request.into_path_query();
-        PathQuery::merge(vec![&balance_query, &revision_query, &all_keys_query])
-            .map_err(Error::GroveDB)
-    }
-
-    /// The query getting all keys and balance and revision
-    pub fn full_identities_query(identity_ids: Vec<[u8; 32]>) -> Result<PathQuery, Error> {
-        let path_queries: Vec<PathQuery> = identity_ids
-            .into_iter()
-            .map(|identity_id| Self::full_identity_query(identity_id))
-            .collect::<Result<Vec<PathQuery>, Error>>()?;
-        PathQuery::merge(path_queries.iter().map(|query| query).collect()).map_err(Error::GroveDB)
-    }
-
-    /// Fetches an identity with all its information from storage.
-    pub fn fetch_proved_full_identity(
-        &self,
-        identity_id: [u8; 32],
-        transaction: TransactionArg,
-    ) -> Result<Option<Vec<u8>>, Error> {
-        let mut drive_operations: Vec<DriveOperation> = vec![];
-        let query = Self::full_identity_query(identity_id)?;
-        let result = self.grove_get_proved_path_query(&query, transaction, &mut drive_operations);
-        match result {
-            Ok(r) => Ok(Some(r)),
-            Err(Error::GroveDB(grovedb::Error::PathKeyNotFound(_)))
-            | Err(Error::GroveDB(grovedb::Error::PathParentLayerNotFound(_)))
-            | Err(Error::GroveDB(grovedb::Error::PathNotFound(_))) => Ok(None),
-            Err(e) => Err(e),
-        }
-    }
-
-    /// Fetches an identity with all its information from storage.
-    pub fn fetch_proved_full_identities(
-        &self,
-        identity_ids: Vec<[u8; 32]>,
-        transaction: TransactionArg,
-    ) -> Result<Option<Vec<u8>>, Error> {
-        let mut drive_operations: Vec<DriveOperation> = vec![];
-        let query = Self::full_identities_query(identity_ids)?;
-        let result = self.grove_get_proved_path_query(&query, transaction, &mut drive_operations);
-        match result {
-            Ok(r) => Ok(Some(r)),
-            Err(Error::GroveDB(grovedb::Error::PathKeyNotFound(_)))
-            | Err(Error::GroveDB(grovedb::Error::PathParentLayerNotFound(_)))
-            | Err(Error::GroveDB(grovedb::Error::PathNotFound(_))) => Ok(None),
-            Err(e) => Err(e),
-        }
-    }
-
-    // /// Fetches an identity with all its information from storage.
-    // pub fn fetch_full_identities(
-    //     &self,
-    //     identity_ids: Vec<[u8; 32]>,
-    //     transaction: TransactionArg,
-    // ) -> Result<BTreeMap<[u8; 32],Identity>, Error> {
-    //     let mut drive_operations: Vec<DriveOperation> = vec![];
-    //     let query = Self::full_identities_query(identity_ids)?;
-    //     let result = self.grove_get_path_query_with_optional(&query, transaction, &mut drive_operations)?;
-    //     result.into_iter().map(|(_, key, element)| {
-    //
-    //     }).collect()
-    // }
-
-    // /// Fetches identities with all its information from storage.
-    // pub fn fetch_full_identities_operations(
-    //     &self,
-    //     identity_ids: Vec<[u8; 32]>,
-    //     transaction: TransactionArg,
-    //     drive_operations: &mut Vec<DriveOperation>,
-    // ) -> Result<BTreeMap<[u8; 32], Option<Identity>>, Error> {
-    //     let mut drive_operations: Vec<DriveOperation> = vec![];
-    //     let query = Self::full_identities_query(identity_ids)?;
-    //     let result = self.grove_get_path_query_with_optional(&query, transaction, &mut drive_operations)?;
-    //     result.into_iter().map(|(_, key, element)| {
-    //
-    //     }).collect()
-    // }
-
-    /// Fetches an identity with all its information from storage.
-    pub fn fetch_full_identity(
-        &self,
-        identity_id: [u8; 32],
-        transaction: TransactionArg,
-    ) -> Result<Option<Identity>, Error> {
-        let mut drive_operations: Vec<DriveOperation> = vec![];
-        self.fetch_full_identity_operations(identity_id, transaction, &mut drive_operations)
-    }
-
-    /// Given an identity, fetches the identity with its flags from storage.
-    pub fn fetch_full_identity_operations(
-        &self,
-        identity_id: [u8; 32],
-        transaction: TransactionArg,
-        drive_operations: &mut Vec<DriveOperation>,
-    ) -> Result<Option<Identity>, Error> {
-        // let's start by getting the balance
-        let balance = self.fetch_identity_balance_operations(
-            identity_id,
-            true,
-            transaction,
-            drive_operations,
-        )?;
-        if balance.is_none() {
-            return Ok(None);
-        }
-        let balance = balance.unwrap();
-        let revision = self
-            .fetch_identity_revision_operations(identity_id, true, transaction, drive_operations)?
-            .ok_or(Error::Drive(DriveError::CorruptedDriveState(
-                "revision not found on identity".to_string(),
-            )))?;
-
-        let public_keys =
-            self.fetch_all_identity_keys_operations(identity_id, transaction, drive_operations)?;
-        Ok(Some(Identity {
-            protocol_version: PROTOCOL_VERSION,
-            id: Identifier::new(identity_id),
-            public_keys,
-            balance,
-            revision,
-            asset_lock_proof: None,
-            metadata: None,
-        }))
     }
 
     /// Given a vector of identities, fetches the identities from storage.
