+++ packages/rs-drive/src/drive/identity/fetch/revision.rs
@@ -1,0 +1,97 @@
+use crate::drive::block_info::BlockInfo;
+use crate::drive::grove_operations::DirectQueryType;
+use crate::drive::grove_operations::QueryTarget::QueryTargetValue;
+use crate::drive::identity::identity_path;
+use crate::drive::identity::IdentityRootStructure::IdentityTreeRevision;
+use crate::drive::Drive;
+use crate::error::drive::DriveError;
+use crate::error::Error;
+use crate::fee::calculate_fee;
+use crate::fee::op::DriveOperation;
+use crate::fee::result::FeeResult;
+use grovedb::Element::Item;
+use grovedb::TransactionArg;
+
+impl Drive {
+    /// Fetches the Identity's revision from the backing store
+    /// Passing apply as false get the estimated cost instead
+    pub fn fetch_identity_revision(
+        &self,
+        identity_id: [u8; 32],
+        apply: bool,
+        transaction: TransactionArg,
+    ) -> Result<Option<u64>, Error> {
+        let mut drive_operations: Vec<DriveOperation> = vec![];
+        self.fetch_identity_revision_operations(
+            identity_id,
+            apply,
+            transaction,
+            &mut drive_operations,
+        )
+    }
+
+    /// Fetches the Identity's revision from the backing store
+    /// Passing apply as false get the estimated cost instead
+    pub fn fetch_identity_revision_with_fees(
+        &self,
+        identity_id: [u8; 32],
+        block_info: &BlockInfo,
+        apply: bool,
+        transaction: TransactionArg,
+    ) -> Result<(Option<u64>, FeeResult), Error> {
+        let mut drive_operations: Vec<DriveOperation> = vec![];
+        let value = self.fetch_identity_revision_operations(
+            identity_id,
+            apply,
+            transaction,
+            &mut drive_operations,
+        )?;
+        let fees = calculate_fee(None, Some(drive_operations), &block_info.epoch)?;
+        Ok((value, fees))
+    }
+
+    /// Creates the operations to get Identity's revision from the backing store
+    /// This gets operations based on apply flag (stateful vs stateless)
+    pub(crate) fn fetch_identity_revision_operations(
+        &self,
+        identity_id: [u8; 32],
+        apply: bool,
+        transaction: TransactionArg,
+        drive_operations: &mut Vec<DriveOperation>,
+    ) -> Result<Option<u64>, Error> {
+        let direct_query_type = if apply {
+            DirectQueryType::StatefulDirectQuery
+        } else {
+            DirectQueryType::StatelessDirectQuery {
+                in_tree_using_sums: false,
+                query_target: QueryTargetValue(1),
+            }
+        };
+        let identity_path = identity_path(identity_id.as_slice());
+        match self.grove_get_raw(
+            identity_path,
+            &[IdentityTreeRevision as u8],
+            direct_query_type,
+            transaction,
+            drive_operations,
+        ) {
+            Ok(Some(Item(encoded_revision, _))) => {
+                let revision = u64::from_be_bytes(encoded_revision.try_into().map_err(|_| {
+                    Error::Drive(DriveError::CorruptedElementType(
+                        "identity revision was not 8 bytes as expected",
+                    ))
+                })?);
+
+                Ok(Some(revision))
+            }
+
+            Ok(None) => Ok(None),
+
+            Ok(Some(..)) => Err(Error::Drive(DriveError::CorruptedElementType(
+                "identity revision was present but was not identified as an item",
+            ))),
+
+            Err(e) => Err(e),
+        }
+    }
+}
